apps/backend/requirements/common.txt
alembic==1.16.2
asyncpg==0.30.0
bcrypt==4.0.1
fastapi==0.115.12
fastapi_mail==1.4.1
greenlet==3.2.3
passlib==1.7.4
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
pydantic-settings==2.9.1
pydantic==2.11.3
pydantic_extra_types==2.10.5
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.2.0
SQLAlchemy==2.0.41
sqladmin[full]==0.21.0
textcase==0.4.3
typer==0.16.0
uvicorn==0.34.2
